+++ src/api/token.js
@@ -1,1 +1,1 @@
-export const token = 'AQBVtKNC17U9g692pbmFB9n9LdipfPZRpaX1iSDujmiVkIVPI_dKFymTfRc_w6dUFIL9wxOcWAbNWY6N0V-ljNkav3UK-F5YyKD68pZ8XE1OG0aKo-e-fahJDKxG_gSMNGGSDg';
+export const token = 'AQD6-mMQakjT78EQXD_-p6nZE1awOIYCtyEDf5mr4oYsf6MOuBC3uLtMXeL0d35a_i6Ttm88ctqw9cK8_euUZiBqaNnFZXuSUGpzAmGVgUDRaY5E4ahoF33EzmZ77P8bIcn5DA';
